+++ src/elements/Block.php
@@ -174,12 +174,6 @@ class Block extends Element implements BlockElementInterface
     public $dirty = false;
 
     /**
-     * @var bool|null Whether this block has been modified.
-     * @deprecated in 2.11.0
-     */
-    private $_modified;
-
-    /**
      * @var array|null All blocks belonging to the same field as this one.
      */
     private $_allElements;
@@ -408,28 +402,6 @@ class Block extends Element implements BlockElementInterface
             $cacheKey = "neoblock-$this->id-collapsed";
             $cacheService->delete($cacheKey);
         }
-    }
-
-    /**
-     * Returns whether this block has been modified.
-     *
-     * @return bool|null
-     * @deprecated in 2.11.0
-     */
-    public function getModified()
-    {
-        return $this->_modified;
-    }
-
-    /**
-     * Sets whether this block has been modified.
-     *
-     * @param bool $value
-     * @deprecated in 2.11.0
-     */
-    public function setModified(bool $value = true)
-    {
-        $this->_modified = $value;
     }
 
     /**